Updated build structure
