+++ src/dirmapper/styles/flat_list_style.py
@@ -2,8 +2,6 @@ from typing import List, Tuple
 from dirmapper.styles.base_style import BaseStyle
 
 class FlatListStyle(BaseStyle):
-    def write_structure(self, structure: List[Tuple[str, int]]) -> str:
-        result = []
-        for item_path, level in structure:
-            result.append(item_path)
+    def write_structure(self, structure: List[Tuple[str, int, str]]) -> str:
+        result = [item_path for item_path, _, _ in structure]
         return '\n'.join(result)
