+++ src/dirmapper/styles/html_style.py
@@ -3,20 +3,21 @@ import os
 from dirmapper.styles.base_style import BaseStyle
 
 class HTMLStyle(BaseStyle):
-    def write_structure(self, structure: List[Tuple[str, int]]) -> str:
+    def write_structure(self, structure: List[Tuple[str, int, str]]) -> str:
         result = ['<ul>']
         previous_level = -1
 
-        for item_path, level in structure:
+        for item_path, level, item in structure:
             if level > previous_level:
                 result.append('<ul>' * (level - previous_level))
             elif level < previous_level:
                 result.append('</ul>' * (previous_level - level))
-            
+
+            relative_path = os.path.relpath(item_path)
             if os.path.isdir(item_path):
-                result.append(f'<li>{os.path.basename(item_path)}/</li>')
+                result.append(f'<li><a href="{relative_path}">{item}/</a></li>')
             else:
-                result.append(f'<li>{os.path.basename(item_path)}</li>')
+                result.append(f'<li><a href="{relative_path}">{item}</a></li>')
 
             previous_level = level
